refactor: reorganize codebase into feature-based modules
- Move JavaScript files from root to organized modules:
  - Core utilities and UI → modules/core/
  - Language learning features → modules/language/
  - Video processing features → modules/video/
- Update import paths in popup.js to use new module structure
- Add comprehensive module organization documentation to README
- Improve code maintainability and scalability with clear separation
