+++ asking-expert/modules/core/index.js
@@ -1,0 +1,7 @@
+/**
+ * Core Module - Entry Point
+ * Exports all core functionality
+ */
+
+export { initializeUI } from './ui.js';
+export * from './utils.js';
