+++ asking-expert/modules/video/index.js
@@ -1,0 +1,7 @@
+/**
+ * Video Module - Entry Point
+ * Exports all video-related functionality
+ */
+
+export { handleExtractQuestions, autoDetectVideoId, handleVideoIdAction } from './video-extractor.js';
+export { handleQuizGeneratorClick } from './youtube-quiz-generator.js';
